# Improve the e2tomoseg_convnet.py and make training faster